requirements.txt
Flask==3.1.2
Flask-WTF==1.2.2
Werkzeug==3.1.3
python-dotenv==1.1.1
Pillow==11.3.0
Flask-SocketIO==5.5.1
python-socketio==5.13.0
python-engineio==4.12.2 